feat: rename Order entity package and remove unused methods
